contributors/i18n/submitted/css3-syntax/i18n-css-syntax-bytes-017.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="ISO-8859-1"/>
<title>CSS3 Syntax, input byte stream: http vs. bom</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='http://www.w3.org/TR/css3-syntax/#input-byte-stream'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='http dom'>
<meta name="assert" content="A UTF-8 signature overrides an HTTP encoding declaration for a stylesheet .">
<style type='text/css'>
.test div { width:50px; }

p { color: gray; margin-top: 3em; font-size: 80%; }
</style>
<link rel="stylesheet" type="text/css" href="support/http1-bom.css">
</head>
<body>



<div class="test"><div id="box" class="эди">&#xA0;</div></div>


<!--Notes:
This test is only valid if the user agent recognises the HTTP and BOM declarations.

In the style sheet, the CSS selector identifies a class name with the following byte sequence: c3 bd c3 a4 c3 a8.  Read as UTF-8 this gives the characters эди.  Read as ISO 8859-1, this gives ГЅГ¤ГЁ.

This HTML page is served as ISO 8859-1 and uses a different byte sequence for the class name (fd e4 e8) that equates to the character sequence эди.  

The class ids will therefore only match if the CSS is read as UTF-8.

The style sheet has a UTF-8 signature at the start, but the HTTP header for the style sheet says ISO 8859-1. If the HTTP header overrides the @charset rule, the CSS will not be read as UTF-8, the selector will not match the class name, and the test will fail.

-->
<script>
test(function() {
assert_equals(document.getElementById('box').offsetWidth, 100);
}, " ");
</script>

<div id='log'></div>

<p class="эди"><span><span id="debug">Text as read from the style sheet: </span><br />(If blank, the style sheet hasn't been read at all.)</span></p>
</body>
</html>
